refactor: improve validation and error handling

+++ domain/service/taskservice.go
@@ -1,11 +1,15 @@
 package service
 
 import (
+	"database/sql"
+	"errors"
+	"fmt"
+
 	"github.com/gin-gonic/gin"
-	"github.com/go-playground/validator/v10"
 	"github.com/muhammad-reda/go-api-gin/domain/entity"
 	"github.com/muhammad-reda/go-api-gin/domain/repository"
 	dto "github.com/muhammad-reda/go-api-gin/dto/task"
+	validation "github.com/muhammad-reda/go-api-gin/validation/body"
 )
 
 type Taskservice interface {
@@ -16,13 +20,23 @@ type Taskservice interface {
 	Delete(ctx *gin.Context, id int64) error
 }
 
-type TaskServiceImplementation struct {
-	taskRepo repository.TaskRepository
+type ErrTaskService struct {
+	Reason string `json:"reason"`
 }
 
-func NewTaskService(taskrepo repository.TaskRepository) Taskservice {
+func (et *ErrTaskService) Error() string {
+	return fmt.Sprintf("Reason: %s", et.Reason)
+}
+
+type TaskServiceImplementation struct {
+	taskRepo repository.TaskRepository
+	userRepo repository.UserRepository
+}
+
+func NewTaskService(taskrepo repository.TaskRepository, userrepo repository.UserRepository) Taskservice {
 	return &TaskServiceImplementation{
 		taskRepo: taskrepo,
+		userRepo: userrepo,
 	}
 }
 
@@ -37,15 +51,18 @@ func (ts *TaskServiceImplementation) GetById(ctx *gin.Context, id int64) (entity
 func (ts *TaskServiceImplementation) Create(ctx *gin.Context) (*entity.Task, error) {
 	var input dto.TaskCreate
 
-	errBindJson := ctx.ShouldBindJSON(&input)
-	if errBindJson != nil {
-		return nil, errBindJson
+	if veer := validation.BodyValidation(ctx, &input); veer != nil {
+		return nil, veer
 	}
 
-	validator := validator.New()
-	errValidate := validator.Struct(input)
-	if errValidate != nil {
-		return nil, errValidate
+	_, errFindUser := ts.userRepo.FindById(ctx, input.UserId)
+	if errFindUser != nil {
+		if errFindUser == sql.ErrNoRows {
+			return nil, &ErrTaskService{
+				Reason: "user not found",
+			}
+		}
+		return nil, errFindUser
 	}
 
 	task := entity.Task{
@@ -65,16 +82,39 @@ func (ts *TaskServiceImplementation) Create(ctx *gin.Context) (*entity.Task, err
 
 func (ts *TaskServiceImplementation) Update(ctx *gin.Context, id int64) (*entity.Task, error) {
 	var input dto.TaskUpdate
-
-	errBindJson := ctx.ShouldBindJSON(&input)
-	if errBindJson != nil {
-		return nil, errBindJson
+	FoundTask, errFindTask := ts.taskRepo.FindById(ctx, id)
+	if errFindTask != nil {
+		if errFindTask == sql.ErrNoRows {
+			return nil, &ErrTaskService{
+				Reason: "task not found",
+			}
+		}
+		return nil, errFindTask
 	}
 
-	validator := validator.New()
-	errValidate := validator.Struct(input)
-	if errValidate != nil {
-		return nil, errValidate
+	if input.Name == "" {
+		input.Name = FoundTask.Name
+	}
+	if input.Description == "" {
+		input.Description = FoundTask.Description
+	}
+	if input.Status == "" {
+		input.Status = FoundTask.Status
+	}
+	if input.UserId == 0 {
+		input.UserId = FoundTask.UserId
+	}
+
+	if veer := validation.BodyValidation(ctx, &input); veer != nil {
+		return nil, veer
+	}
+
+	_, errFindUser := ts.userRepo.FindById(ctx, input.UserId)
+	if errFindUser != nil {
+		if errFindUser == sql.ErrNoRows {
+			return nil, errors.New("user not found")
+		}
+		return nil, errFindUser
 	}
 
 	task := entity.Task{
@@ -93,6 +133,15 @@ func (ts *TaskServiceImplementation) Update(ctx *gin.Context, id int64) (*entity
 }
 
 func (ts *TaskServiceImplementation) Delete(ctx *gin.Context, id int64) error {
+	_, errFindTask := ts.taskRepo.FindById(ctx, id)
+	if errFindTask != nil {
+		if errFindTask == sql.ErrNoRows {
+			return &ErrTaskService{
+				Reason: "task not found",
+			}
+		}
+		return errFindTask
+	}
 	errDelete := ts.taskRepo.Delete(ctx, id)
 	return errDelete
 }
